+++ src/plugins/core/execute/index.ts
@@ -1,5 +1,6 @@
 export * from './createRequestInterceptor';
 export * from './lazyVariableInterceptor';
 export * from './logResponseInterceptor';
+export * from './metaDataNameInterceptor';
 export * from './processedHttpRegionInterceptor';
 export * from './regionScopedVariablesInterceptor';