+++ src/plugins/core/response/index.ts
@@ -1,3 +1,2 @@
-export * from './handleNameMetaData';
 export * from './jsonResponseInterceptor';
 export * from './setLastResponseInVariables';